requirements.txt
firebase_admin == 6.1.0
line-bot-sdk == 2.4.2
